AcneSkinCare/acne-ai-skin-digital-twin/requirements.txt
Streamlit==1.10.0
opencv-python==4.5.3.20210927
mediapipe==0.8.9
torch==1.10.0
tensorflow==2.7.0
numpy==1.21.2
pandas==1.3.3
matplotlib==3.4.3
scikit-learn==0.24.2
Pillow==8.4.0
streamlit-authenticator==0.1.0